Send Return Contacts to Bridge.
Including thumbnailImageData and imageData.

We'll have to see how performance is affected with returning the imageData as a Base64.

Might have to just use the thumbnail and then if I do need a big picture, I can do a separate fetch on an individual Contact record for that. We'll see.