Move to string format instead of object format for addresses to match outpout
Currently, the library outputs an email and a phone with a “label” and
“value” object.

However, the library outputs a postal address with a “label” and then
several values (street, city, etc) as siblings to the label instead of
it’s own “value” object.

this means that saving back as a “Label” and “Value” object only works
on initial save but causes a crash if re-saving into a contact with
more than postal address.

So I’ve changed the format of my PR to allow for saving and updating of
addresses in this format:

postalAddress = [
{
label: “home”
city: “Toronto”,
country: “Canada”
},
{
label: “Work”,
city: “New York City”,
Country: “United States”
}
]

This is instead of :

postalAddress = [
{
label: “Work”,
value: {
         city: “New York City”,
         Country: “United States”
        }
}
]

This follows the output of the library and so doesn’t rock the boat.