Test Promise usage on `userCanAccessContactsAsPromise`.
This will easy to migrate from Callbacks to Promises. But, it’s more difficult to support both so we’ll need to do this in one fell swoop.

I created this duplicate function to test out the Promise implementation. Works great. So I’m going to keep it for the Sample app. Any Promises we’ll just use `AsPromise` for now.