`Contacts.requestAccessToContactsAsPromise` Implemented.
This returns a Promise and will eventually replace `Contacts.requestAccessToContacts`, which returns a Callback (but doesn’t actually in this version).

We might consider using React Native’s PermissionsAndroid module to handle permission checking because this requires an additional step of adding `onRequestPermissionsResult` method in the MainActivity class. You can see this in the ExampleApp.

I *believe* if we use React Native’s PermissionsAndroid we won’t need to do this. But that involves restructuring how we return the module. We’d need to modify `index.js` and handle logic in there based on the platform, etc. Easy enough but requires some re-write.