Add `alreadyRequestedAccessToContacts` to Android.
Fixes #73.

This is an interesting one, because Android handles things a little differently than iOS.

With Android, you can request for the same permission again if the User deny’s the first time. The second request will show the same dialog box but will also have a checkbox that says “Don’t ask again.”

If the user denies the second time (or any subsequent time) and has this “Don’t ask again” checked then no further dialog boxes will appear when requesting that permission.

However, nothing in the Android SDK really tells us that they have checked this. If we request that permission again it will just come back as denied, just like any other time. So we need to handle it differently.

What we’re doing now is just storing a SharedPreference called `ALREADY_REQUESTED_ACCESS_TO_CONTACTS` that is false at the beginning and on the first request for permission, it gets set to true.

Thinking about this further, what we’re really after is whether or not we need to provide some extra instructions to the user or open up the Permissions/Privacy settings of the app so they can manually change the permissions themselves.

I think for a new release, we should get rid of `alreadyRequestedAccessToContacts` and replace it with something like `permanentlyDeniedAccessToContacts`.

This method would return true on iOS after the User has denied the permission request just once (because that’s all it takes on iOS) and will return true on Android only when the User has denied permission and checked off the “Don’t ask again” checkbox.